+++ packages/typespec-autorest/test/multipart.test.ts
@@ -7,7 +7,7 @@ it("model properties are spread into individual parameters", async () => {
     `
     model Form { name: HttpPart<string>, profileImage: HttpPart<bytes> }
     op upload(@header contentType: "multipart/form-data", @multipartBody body: Form): void;
-    `
+    `,
   );
   const op = res.paths["/"].post;
   deepStrictEqual(op.parameters, [
@@ -30,7 +30,7 @@ it("part of type `bytes` produce `type: file`", async () => {
   const res = await openApiFor(
     `
     op upload(@header contentType: "multipart/form-data", @multipartBody body: { profileImage: HttpPart<bytes> }): void;
-    `
+    `,
   );
   const op = res.paths["/"].post;
   deepStrictEqual(op.parameters, [
@@ -47,7 +47,7 @@ it("part of type `bytes[]` produce `type: array, items: { type: string, format: 
   const res = await openApiFor(
     `
     op upload(@header contentType: "multipart/form-data", @multipartBody _: { profileImage: HttpPart<bytes>[]}): void;
-    `
+    `,
   );
   const op = res.paths["/"].post;
   deepStrictEqual(op.parameters, [
@@ -68,7 +68,7 @@ it("part of type `string` produce `type: string`", async () => {
   const res = await openApiFor(
     `
     op upload(@header contentType: "multipart/form-data", @multipartBody body: { name: HttpPart<string> }): void;
-    `
+    `,
   );
   const op = res.paths["/"].post;
   deepStrictEqual(op.parameters, [
@@ -87,7 +87,7 @@ it("part of type `object` produce `type: string`", async () => {
     `
     #suppress "@azure-tools/typespec-autorest/unsupported-multipart-type" "For test"
     op upload(@header contentType: "multipart/form-data", @multipartBody _: { address: HttpPart<{city: string, street: string}>}): void;
-    `
+    `,
   );
   const op = res.paths["/"].post;
   deepStrictEqual(op.parameters, [
@@ -105,7 +105,7 @@ describe("legacy implicit form", () => {
     const res = await openApiFor(
       `
       op upload(@header contentType: "multipart/form-data", profileImage: bytes): void;
-      `
+      `,
     );
     const op = res.paths["/"].post;
     deepStrictEqual(op.parameters, [
@@ -122,7 +122,7 @@ describe("legacy implicit form", () => {
     const res = await openApiFor(
       `
       op upload(@header contentType: "multipart/form-data", @doc("Part doc") profileImage: bytes): void;
-      `
+      `,
     );
     const op = res.paths["/"].post;
     deepStrictEqual(op.parameters, [
@@ -140,7 +140,7 @@ describe("legacy implicit form", () => {
     const res = await openApiFor(
       `
       op upload(@header contentType: "multipart/form-data", profileImage: bytes[]): void;
-      `
+      `,
     );
     const op = res.paths["/"].post;
     deepStrictEqual(op.parameters, [
@@ -161,7 +161,7 @@ describe("legacy implicit form", () => {
     const res = await openApiFor(
       `
       op upload(@header contentType: "multipart/form-data", name: string): void;
-      `
+      `,
     );
     const op = res.paths["/"].post;
     deepStrictEqual(op.parameters, [
@@ -180,7 +180,7 @@ describe("legacy implicit form", () => {
       `
       #suppress "@azure-tools/typespec-autorest/unsupported-multipart-type" "For test"
       op upload(@header contentType: "multipart/form-data", address: {city: string, street: string}): void;
-      `
+      `,
     );
     const op = res.paths["/"].post;
     deepStrictEqual(op.parameters, [