+++ packages/typespec-azure-core/test/types/arm-resource-identifier.test.ts
@@ -41,7 +41,7 @@ describe("when used as ref", () => {
 
   it("use with single type and scopes", async () => {
     const type = await compileAsRef(
-      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}]>`
+      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}]>`,
     );
     expect(getArmResourceIdentifierConfig(runner.program, type)).toEqual({
       allowedResources: [
@@ -55,7 +55,7 @@ describe("when used as ref", () => {
 
   it("use multiple single type and scopes", async () => {
     const type = await compileAsRef(
-      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}, {type:"Microsoft.RP/type2", scopes:["tenant", "resourceGroup"]}]>`
+      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}, {type:"Microsoft.RP/type2", scopes:["tenant", "resourceGroup"]}]>`,
     );
     expect(getArmResourceIdentifierConfig(runner.program, type)).toEqual({
       allowedResources: [
@@ -96,7 +96,7 @@ describe("when used as scalar extends", () => {
 
   it("use with single type and scopes", async () => {
     const type = await compileOnScalar(
-      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}]>`
+      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}]>`,
     );
     expect(getArmResourceIdentifierConfig(runner.program, type)).toEqual({
       allowedResources: [
@@ -110,7 +110,7 @@ describe("when used as scalar extends", () => {
 
   it("use multiple single type and scopes", async () => {
     const type = await compileOnScalar(
-      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}, {type:"Microsoft.RP/type2", scopes:["tenant", "resourceGroup"]}]>`
+      `armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}, {type:"Microsoft.RP/type2", scopes:["tenant", "resourceGroup"]}]>`,
     );
     expect(getArmResourceIdentifierConfig(runner.program, type)).toEqual({
       allowedResources: [